\name{candies_levels}
\alias{candies_levels}
\docType{data}
\title{
Chocolate Candy Levels
}
\description{
Levels of all chocolate candy's attributes
}
\usage{data(candies_levels)}
\format{
  A data frame with 6 observations on the following variable.
  \describe{
	2 + 2 + 2 = 6
    Dark, Hard, Milk, No Nuts, Nuts}  
}
\details{
SAS Technical Report R-109 was created and written by Warren F. Kuhfeld. Development and support of the TRANSREJG procedure is the responsibility of Warren F. Kuhfeld.
available at : http://support.sas.com/documentation/onlinedoc/v82/techreport_r109.pdf}
\source{
SAS
}
\references{
SAS Technical Report R-l 09
}
\examples{
library(faisalconjoint)
data(candies_levels)
candies_levels
}

