\name{carpet_levels}
\alias{carpet_levels}
\docType{data}
\title{
Carpet Levels
}
\description{
Levels of all carpet's attributes
}
\usage{data(carpet_levels)}
\format{
  A data frame with 13 observations on the following variable.
  \describe{ 
3 + 3 + 3 + 2 + 2 = 13}
A, B, C, K2R, Glory, Bissell, $1.19, $1.39, $1.59, No, Yes, No, Yes
  }
\details{
http://www.unileon.es/ficheros/servicios/informatica/spss/english/IBM-SPSS_conjoint.pdf
}
\source{
SPSS 19
}
\references{
Green, P. E. and Y. Wind (1973), Multi-Attribute Decisions in Marketing. New York: Holt, Rinehart & Winston
}
\examples{
library(faisalconjoint)
data(carpet_levels)
carpet_levels

}

