\name{cleanser_levels}
\alias{cleanser_levels}
\docType{data}
\title{
Cleanser Levels
}
\description{
Levels of all chocolate cleanser's attributes
}
\usage{data(cleanser_levels)}
\format{
  A data frame with 6 observations on the following variable.
  \describe{ 2 + 2 + 2 = 6
   Liquid, Powder, Phosphate-free, Phosphate-base, HATCO, Generic
  }
}
\details{
Hair et al. 2009, Multivariate Data Analysis, 7/ed, Prenhal
}
\source{
Hair et al. 2009, Multivariate Data Analysis, 7/ed, Prenhal
}
\references{
Hair et al. 2009, Multivariate Data Analysis, 7/ed, Prenhal
}
\examples{
library(faisalconjoint)
data(cleanser_levels)
cleanser_levels
}

