\name{mobile_levels}
\alias{mobile_levels}
\docType{data}
\title{
Mobile Phone Set Data
}
\description{
Data Collected through Sample Survey from Karachi, Pakistan from July 1,2007  by July 31, 2007
}
\usage{data("mobile_levels")}
\format{
  A data frame with 21 observations.

}

\examples{
library(faisalconjoint)
data(mobile_levels)
mobile_levels
}
\keyword{datasets}