\name{traditional_levels}
\alias{traditional_levels}
\docType{data}
\title{
Traditional Conjoint Analysis
}
\description{
Traditional Conjoint Analysis with Excel
}
\usage{data(traditional_levels)}
\format{
  A data frame with 8 observations on the following variable.
  \describe{
    A,B,C, Red, Blue, 50, 100, 150
  }
}
\details{
http://www.sawtoothsoftware.com/download/techpap/caexcel.pdf
}

\source{
Orme, B. (2010) Getting Started with Conjoint Analysis: Strategies for Product Design and Pricing Research. Second Edition, Madison, Wis.: Research Publishers LLC.
}
\references{
{Orme, B. (2010) Getting Started with Conjoint Analysis: Strategies for Product Design and Pricing Research. Second Edition, Madison, Wis.: Research Publishers LLC.}
}

\examples{
library(faisalconjoint)
data(traditional_levels)
traditional_levels
}
\keyword{datasets}
